Fix reading descriptor files from directory in the project
On the new model creation wizard the "Model" entity was not being selected as root. Fixed.
The outer catch was being used to handle the CustomService. This catch is for all error so
a specific catch was added.
CustomServices will trigger a reparse when the first machine is added. It will find the descriptors location
by scanning all projects inside the workspace.